DEV-1752: dedupe fixtures via import, not a shared module
sonar.cpd.exclusions did not clear the new-code duplication gate for the
shared fixtures file. Drop that approach: delete tests/_mode_a_scope_fixtures.py,
restore test_dev1745_mode_a_door.py's original inline fixtures (its diff is now
just the DEV-1752 xfail flip), and import the orders/customers/regions ScopeFrame
fixtures into test_dev1752_subquery_scope.py directly from the DEV-1745 door
pack. No new duplicated code is introduced, so the gate has nothing to flag.
Reverts the sonar-project.properties CPD exclusion.
