+++ sonar-project.properties
@@ -15,14 +15,7 @@ sonar.organization=motleyai
 # expectations, etc.). The duplication is intentional — extracting the
 # shared bits into a helper would tangle dialect concerns without changing
 # the per-dialect specifics. Suppress at the file level.
-#
-# DEV-1752: `tests/_mode_a_scope_fixtures.py` is the shared `orders → customers
-# → regions` ScopeFrame fixture module. That same fixture triple is replicated
-# inline across ~20 legacy test files (test_dev1745_reachability.py, etc.); this
-# module consolidates the two files DEV-1752 touches, but necessarily matches
-# the legacy inline copies. Migrating all of them is out of scope, so exclude
-# the shared source from CPD rather than trip the new-code duplication gate.
-sonar.cpd.exclusions=tests/integration/test_integration_mysql.py,tests/integration/test_integration_clickhouse.py,tests/integration/test_integration_sqlserver.py,tests/_mode_a_scope_fixtures.py
+sonar.cpd.exclusions=tests/integration/test_integration_mysql.py,tests/integration/test_integration_clickhouse.py,tests/integration/test_integration_sqlserver.py
 
 # DEV-1730: the Cube-JS import fixtures under tests/fixtures/cube_js/ are Cube
 # FILTER_PARAMS DSL, not executable JavaScript. SonarJS misreads the
